Crate rv32i

RISC-V 32bit.

This example implements a minimal rv32i userspace emulator with an interpreter and jit compiler to demonstrate the juicebox crate.

The emulator only implements a very limited syscall surface, sufficient to run the example software in examples/rv32i-guest/. However, enough to run a multi-threaded example with TLS support.

It aims at simplicity rather than being highly optimized and uses unwraps throughout the implementation to crash the emulator on unexpected behavior.

Self-modifying code is not supported and writes to executable memory crash the simulator. In theory the interpreter actually has no problem with self-modifying code as there is no decode cache or anything alike. For the jit one could achieve support by invalidating the translation cache on writes to executable memory and thinking about whether to handle those invalidations as precise or imprecise exceptions in a translation block.

Structs

Btype

GuestState

The rv32i guest state.

Itype

JitRet

Jit function return value.

Jtype

PhdrInfo

ProcessCtx

A process context which serves as container for multiple threads and maintains shared information.

Rtype

Stype

ThreadCtx

Guest thread context that can be saved and restored to achieve scheduling multiple guest threads on a single guest vcpu.

Utype

Enums

ExitReason

Insn

ThreadState

Guest thread states.

Constants

A0

Register ABI name - Argument 0.

A1

Register ABI name - Argument 1.

A2

Register ABI name - Argument 2.

A3

Register ABI name - Argument 3.

A4

Register ABI name - Argument 4.

A7

Register ABI name - Argument 7.

JIT_EBREAK

Jit exit when the guest is about to execute an ebreak instruction.

JIT_ECALL

Jit exit when the guest is about to execute an ecall instruction.

JIT_LD_FAULT

Jit exit when a load instruction faults due to an out of bounds access or read protection violation.

JIT_ST_FAULT

Jit exit when a store instruction faults due to an out of bounds access or write protection violation.

JIT_TB_END

Jit exit when the end of a translation block is reached.

PROT_R

Protection flag - read access.

PROT_W

Protection flag - write access.

PROT_X

Protection flag - execute access.

SP

Register ABI name - Stack Pointer.

TP

Register ABI name - Thread Pointer.

Functions

decode

Decode the riscv instruction bytes insn into an Insn.

load_elf

Load the elf file into the guest and return the entry point and information about the program header.

setup_main_stack

Map zeroed out stack of size bytes at the end of the guest address space and return (final stack pointer, stack bottom). The stack is initialized according to the SystemV abi (Process Initialization - Stack State).

Type Aliases

JitFn

Function signature for a jit compiled translation block.