rv32i

Function setup_main_stack

pub fn setup_main_stack(
    guest: &mut GuestState,
    size: u32,
    prog: &[u8],
    phdr: PhdrInfo,
) -> (u32, u32)

Map zeroed out stack of size bytes at the end of the guest address space and return (final stack pointer, stack bottom). The stack is initialized according to the SystemV abi (Process Initialization - Stack State).

https://github.com/johannst/dynld/tree/main/02_process_init https://github.com/torvalds/linux/blob/c84d3e3130dfe1058cb27dc78e7ad8bd36f0545a/fs/binfmt_elf.c#L165