rv32i

Struct GuestState

pub struct GuestState { /* private fields */ }

The rv32i guest state.

Implementations

impl GuestState

pub fn new(mem_size: usize) -> Self

Create a new guest state with the virtual address space [0..mem_size). The virtual address space is initially unmapped, and accessing it from the guest will raise a fault.

pub fn read_reg(&self, r: u32) -> u32

Read the guest register r.

pub fn write_reg(&mut self, r: u32, val: u32)

Write val to the guest register r.

pub fn check_prot(&self, addr: u32, len: usize, prot: u8) -> bool

Check if the address range [addr..addr+len) has at least the prot protection flag set.

pub fn set_prot(&mut self, addr: u32, len: usize, prot: u8)

Set the prot protection flag for the address range [addr..addr+len). This will just overwrite the current protection flags, and does not check if there were other protection flags set.

pub fn map_mem(&mut self, addr: u32, data: &[u8], prot: u8)

Map the virtual address range [addr..addr+data.len) with the protection provided in prot and initialize the memory with data.

pub fn map_mem_zero(&mut self, addr: u32, len: usize, prot: u8)

Map the virtual address range [addr..addr+len) with the protection provided in prot and initialize the memory with 0.

pub fn write_mem(&mut self, addr: u32, data: &[u8])

Write data to the virtual address range [addr..addr+data.len). This performs a check if the address range has the write protection set.

pub fn read_mem(&self, addr: u32, data: &mut [u8])

Read from the virtual address range [addr..addr+data.len) into data. This performs a check if the address range has the read protection set.

pub fn slice_mem(&self, addr: u32, len: usize) -> &[u8]

Get a slice for the virtual address range [addr..addr+len). This performs a check if the address range has the read protection set.

pub fn fetch_insn(&self, pc: u32) -> u32

Fetch an instruction from pc. This performs a check if the address is properly aligned and the range has the exec protection set.

pub fn interpret(&mut self) -> ExitReason

Execute guest on interpreter until an ExitReason is hit, starting from self.pc.

pub fn interpret_insn(&mut self, insn: Insn) -> Result<Option<u32>, ExitReason>

Interpret a single guest instruction insn.

pub fn jit(&mut self) -> ExitReason

Execute guest from jit compiled code until an ExitReason is hit, starting from self.pc.

pub fn translate_next_block(&mut self) -> JitFn

Translate the next block of guest code at self.pc and return a JitFn to the compiled TB.